Have a line and disk version of ActSrcOp set up and compiling.
