+++ amr-wind/wind_energy/actuator/ActuatorModel.H
@@ -4,7 +4,8 @@
 #include "amr-wind/core/Factory.H"
 #include "amr-wind/wind_energy/actuator/actuator_types.H"
 #include "amr-wind/wind_energy/actuator/actuator_ops.H"
-#include "amr-wind/wind_energy/actuator/ActSrcOp.H"
+#include "amr-wind/wind_energy/actuator/ActSrcLineOp.H"
+#include "amr-wind/wind_energy/actuator/ActSrcDiskOp.H"
 #include "amr-wind/wind_energy/actuator/ActParser.H"
 
 namespace amr_wind {
